Updates approach for epoch handling in rpm comparisons (#438)
* Updates approach for epoch handling in rpm comparisons to ignore epochs
if not explicitly available from both sides of comparison. Fixes #437

This approach adds -1 as epoch value in the struct to identify
"not-specified" rather than defaulting to 0, per rpm spec, so that the
comparison logic can identify when it is provided vs missing.

During the comparison if both sides to not have an explicitly set epoch
it will skip the epoch check as unreliable and compare the remaining
components. This is done to handle messy data in RedHat vuln feeds where
often the sourceRpm versions do not include epochs when they should, and
defaulting to zero or using the epoch of the binary version is also
incorrect.

Signed-off-by: Zach Hill <zach@anchore.com>

* Uses switch instead of if-else-chain per linter suggestion

Signed-off-by: Zach Hill <zach@anchore.com>

* Fix import order/format for linter

Signed-off-by: Zach Hill <zach@anchore.com>

* Comment out strictCompare for linter

Signed-off-by: Zach Hill <zach@anchore.com>

* Minor cleanup

Signed-off-by: Dan Luhring <dan.luhring@anchore.com>

* Remove commented out function

Signed-off-by: Dan Luhring <dan.luhring@anchore.com>

* Refactor RPM version comparison to make missing epoch explicit

Signed-off-by: Dan Luhring <dan.luhring@anchore.com>

* More cleanup

Signed-off-by: Dan Luhring <dan.luhring@anchore.com>

* change epoch to pointer type

Signed-off-by: Alex Goodman <alex.goodman@anchore.com>

* add rpmdb matcher tests for explicit epoch being passed

Signed-off-by: Alex Goodman <alex.goodman@anchore.com>

* explicitly pass the epoch on package versions in the rpmdb matcher

Signed-off-by: Alex Goodman <alex.goodman@anchore.com>

Co-authored-by: Dan Luhring <dan.luhring@anchore.com>
Co-authored-by: Alex Goodman <alex.goodman@anchore.com>