Include package type column in table output
This helps avoid confusion between packages of the same name but different types.

I've hit this on a number of occasions, some examples below:
 - `tar` could be either a node package or a linux apk/rpm/deb
 - `msgpack` is a node package but also a python package
 - `jsonpointer` is also a node and/or python package

In each case when I saw the vuln reported I unluckily picked
the "wrong" one and it took some digging to realise the issue
or even that there was another type of package with the same
name at all.

The "type" is a succinct representation of _where_ Grype found
this package which should make things a lot clearer.

Signed-off-by: Tom Sparrow <793763+sparrowt@users.noreply.github.com>