feat: package qualifier for platform CPE
This allows filtering vulnerability matches that are only applicable
when running on specific platforms.  It currently supports filtering
matches that are only applicable for windows, debian, and ubuntu when
the underlying distro is known and does not match.

Additionally, wordpress platform matches are always filtered since
wordpress plugins are not currently discoverable by syft and can be
a significant source of false-positive matches.  These are already
filtered when the target software component of the CPE is used
rather than a running on platform CPE configuration.

Signed-off-by: Weston Steimel <weston.steimel@anchore.com>