chore: log unsupported package qualifier as debug
Logs unsupported package qualifiers at `debug` level rather than
`warning`.  The message is only meant to convey that there are new
qualifiers available in grype-db that the version of grype being used
cannot take advantage of to improve matching behavior; however, the
warning is confusing to users and may make it seem like grype is in a
broken state.

Signed-off-by: Weston Steimel <weston.steimel@anchore.com>