[chore] Add mastodon link to README.md
This follows the same pattern as the other badges at the top of the readme. It adds the mastodon link to the Grype account.

This also means that the link back here from the Mastodon account's profile page will show as 'Validated' once it lands, which gives the account more authenticity.

I also converted all the links to html from markdown format so that I could add the necessary `rel="me"` required by mastodon validation.

Signed-off-by: Alan Pope <alan.pope@anchore.com>