chore: improve release quality gate
The release quality gate step polls to make sure the commit to be
released has a green CI run. However if a PR is merged immediately
before the release job is run, this step often times out. Raise the
timeout to 20 minutes, which covers the time taken by recent CI runs.

Additionally, include the "Quality tests" step from validations.yaml in
the tests to block on. It should always have been included.

Signed-off-by: Will Murphy <willmurphyscode@users.noreply.github.com>