ruby version specific comparator
RubyGems uses a version format similar to but not identical to semver.
Previously, Grype used semver to compare the versions of ruby gems,
which is usually correct but occasionally leads to failed comparisions
or incorrect matching, especially with alphabetically labeled
pre-releases. Switch to a new custom version comparator based on the
comparison used by Ruby itself.

Signed-off-by: Will Murphy <willmurphyscode@users.noreply.github.com>