fix: preserve target sw when unknown (#2772)
* fix: preserve target sw when unknown

Normally, when matching CPEs with the target SW filled as a non-wild,
Grype attempts to normalize the target SW field to a known package type,
e.g. ".net" and "dotnet" both compare as "dotnet" at match time.
However, this code dropped unknown target SW types, resulting in false
negatives for target SW entries that did not correspond to a particular
package type, e.g. "chromium". Therefore, if a target SW is not known to
Grype, and not a wildcard, lowercase it and consider it for matching.

Signed-off-by: Will Murphy <willmurphyscode@users.noreply.github.com>

* hyphen is also not a target sw

Signed-off-by: Will Murphy <willmurphyscode@users.noreply.github.com>

---------

Signed-off-by: Will Murphy <willmurphyscode@users.noreply.github.com>