refactor: csv is type of table
Previous draft of this feature included the csv format in its own
present, but that is effectively just a copy of the table presenter with
a different separation strategy. Therefore, combine the csv and table
presenter so that if, for example, a new column needs to be added, it
only needs to be done in one please.

Signed-off-by: Will Murphy <willmurphyscode@users.noreply.github.com>